Migrate to microservices architecture for quiz app
Reorganized the monolithic quiz app into two distinct microservices: `question-service` and `quiz-service`. Introduced service-specific entities, controllers, and repositories, along with inter-service communication utilizing Feign clients. Updated configurations to support Eureka for service discovery and replaced application properties with YAML setup for clearer management.
