+++ question-service/src/main/java/com/dev/yatin/questionservice/QuestionServiceApplication.java
@@ -1,0 +1,13 @@
+package com.dev.yatin.questionservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class QuestionServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(QuestionServiceApplication.class, args);
+    }
+}
