+++ question-service/src/main/java/com/dev/yatin/questionservice/entity/Category.java
@@ -1,0 +1,18 @@
+package com.dev.yatin.questionservice.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Entity
+@Data
+@Table(name = "categories")
+public class Category {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column(unique = true, nullable = false)
+    private String name;
+
+    // Constructor, getters, and setters are handled by @Data
+}
