+++ question-service/src/main/java/com/dev/yatin/questionservice/repository/CategoryRepository.java
@@ -1,0 +1,13 @@
+package com.dev.yatin.questionservice.repository;
+
+import com.dev.yatin.questionservice.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+    Optional<Category> findByName(String name);
+    boolean existsByName(String name);
+}
