+++ question-service/src/main/resources/application.yml
@@ -1,0 +1,24 @@
+server:
+  port: 8081
+
+spring:
+  application:
+    name: question-service
+  datasource:
+    url: jdbc:postgresql://localhost:5432/user_db
+    username: user_name
+    password: user_password
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+    properties:
+      hibernate:
+        format_sql: true
+
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: http://localhost:8761/eureka/
+  instance:
+    preferIpAddress: true
