+++ quiz-service/src/main/java/com/dev/yatin/quizservice/QuizServiceApplication.java
@@ -1,0 +1,15 @@
+package com.dev.yatin.quizservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableFeignClients
+public class QuizServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(QuizServiceApplication.class, args);
+    }
+}
