+++ quiz-service/src/main/java/com/dev/yatin/quizservice/dto/AnswerRequest.java
@@ -1,0 +1,11 @@
+package com.dev.yatin.quizservice.dto;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@RequiredArgsConstructor
+public class AnswerRequest {
+    private String option_selected;
+    private Long question_id;
+}
