+++ quiz-service/src/main/java/com/dev/yatin/quizservice/dto/CategoryDto.java
@@ -1,0 +1,13 @@
+package com.dev.yatin.quizservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CategoryDto {
+    private Integer id;
+    private String name;
+}
