+++ quiz-service/src/main/java/com/dev/yatin/quizservice/dto/Response.java
@@ -1,0 +1,13 @@
+package com.dev.yatin.quizservice.dto;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@Data
+@RequiredArgsConstructor
+public class Response {
+    private List<AnswerRequest> data;
+    private Long quizId;
+}
