--- src/main/java/com/dev/yatin/quizapp/QuizappApplication.java
@@ -1,13 +1,0 @@
-package com.dev.yatin.quizapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class QuizappApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(QuizappApplication.class, args);
-	}
-
-}
