--- src/main/java/com/dev/yatin/quizapp/controller/QuestionController.java
@@ -1,32 +1,0 @@
-package com.dev.yatin.quizapp.controller;
-
-import com.dev.yatin.quizapp.entity.Question;
-import com.dev.yatin.quizapp.service.QuestionService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("question")
-public class QuestionController {
-
-    @Autowired
-    QuestionService questionService;
-
-    @GetMapping("allQuestion")
-    public ResponseEntity<List<Question>> getAllQuestions() {
-        return questionService.getAllQuestions();
-    }
-
-    @GetMapping("category/{category}")
-    public List<Question> getQuestionsByCategory(@PathVariable int category) {
-        return questionService.getQuestionsByCategory(category);
-    }
-
-    @PostMapping("addQuestion")
-    public ResponseEntity<Question> addQuestion(@RequestBody Question question) {
-        return questionService.addQuestion(question);
-    }
-}
