--- src/main/java/com/dev/yatin/quizapp/dao/CategoryDao.java
@@ -1,13 +1,0 @@
-package com.dev.yatin.quizapp.dao;
-
-import com.dev.yatin.quizapp.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface CategoryDao extends JpaRepository<Category, Integer> {
-    Optional<Category> findByName(String name);
-    boolean existsByName(String name);
-}
