--- src/main/java/com/dev/yatin/quizapp/dao/QuizDao.java
@@ -1,10 +1,0 @@
-package com.dev.yatin.quizapp.dao;
-
-import com.dev.yatin.quizapp.entity.Quiz;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuizDao extends JpaRepository<Quiz,Long> {
-
-}
