--- src/main/java/com/dev/yatin/quizapp/entity/AnswerRequest.java
@@ -1,11 +1,0 @@
-package com.dev.yatin.quizapp.entity;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@RequiredArgsConstructor
-public class AnswerRequest {
-    private String option_selected;
-    private Long question_id;
-}
