--- src/main/java/com/dev/yatin/quizapp/entity/Category.java
@@ -1,18 +1,0 @@
-package com.dev.yatin.quizapp.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-@Entity
-@Data
-@Table(name = "categories")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(unique = true, nullable = false)
-    private String name;
-
-    // Constructor, getters, and setters are handled by @Data
-}
