--- src/main/java/com/dev/yatin/quizapp/entity/Question.java
@@ -1,37 +1,0 @@
-package com.dev.yatin.quizapp.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "questions")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Question {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id", nullable = false)
-    private Category category;
-
-    @Column(columnDefinition = "TEXT")
-    private String question;
-
-    @Column(name = "option_a")
-    private String optionA;
-    @Column(name = "option_b")
-    private String optionB;
-    @Column(name = "option_c")
-    private String optionC;
-    @Column(name = "option_d")
-    private String optionD;
-
-    private String correctOption;
-    private String difficulty;
-}
