--- src/main/java/com/dev/yatin/quizapp/entity/Quiz.java
@@ -1,25 +1,0 @@
-package com.dev.yatin.quizapp.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@Table(name = "quiz")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Quiz {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String title;
-
-    @ManyToMany(cascade = CascadeType.ALL)
-    private List<Question> questions;
-
-}
