--- src/test/java/com/dev/yatin/quizapp/QuizappApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dev.yatin.quizapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QuizappApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
